requirements.txt
google-adk>=0.1.0
google-genai>=1.0.0